pyproject.toml
[tool.poetry]
name = "bc-deloitte"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.10.1"
isort = "^5.12.0"
pre-commit = "^3.5.0"
pytest-django = "^4.5.2"
model-bakery = "^1.16.0"
bandit = "^1.7.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
